OllamaTesting/requirements.txt
Flask
langchain
langchain-community
chromadb
langchain-google-genai
google-generativeai
ollama